+++ src/actions/sync/clean_branches.ts
@@ -1,8 +1,8 @@
 import { default as chalk } from 'chalk';
 import prompts from 'prompts';
-import { TContext } from '../lib/context';
-import { KilledError } from '../lib/errors';
-import { deleteBranchAction, isSafeToDelete } from './delete_branch';
+import { TContext } from '../../lib/context';
+import { KilledError } from '../../lib/errors';
+import { deleteBranchAction, isSafeToDelete } from '../delete_branch';
 
 /**
  * This method is assumed to be idempotent -- if a merge conflict interrupts
